fixed a few bugs

+++ SELVoltageMonitor/SELVoltageMonitor.cydsn/filter.h
@@ -15,12 +15,13 @@
     
     #include <stdint.h>
     #include <stdbool.h>
+    #include <stdio.h>
     
     typedef struct {
         uint8_t length; // Filter length
         int16_t * taps; // Array of filter taps
         uint16_t gain; // Gain of filter (for use in calculations)
-        int32_t * samples; // Array of samples
+        uint32_t * samples; // Array of samples
         bool is_charged;// Returns true when filter is charged.
         uint8_t num_samples;
     } fir_filter_t;
@@ -32,7 +33,7 @@
     returns: none
     notes: none
     ***************/
-    void initialize_filter(uint8_t length, int16_t * taps, int16_t gain, int32_t * samples, fir_filter_t * filter);
+    void initialize_filter(uint8_t length, int16_t * taps, uint16_t gain, uint32_t * samples, fir_filter_t * filter);
     
     /**************
     function: insert_filter_value
@@ -41,7 +42,7 @@
     returns: none
     notes: none
     ***************/
-    void insert_filter_value(fir_filter_t *filter, int32_t value);
+    void insert_filter_value(fir_filter_t *filter, uint32_t value);
     
     /**************
     function: get_filtered_value
@@ -50,7 +51,7 @@
     returns: none
     notes: none
     ***************/
-    int16_t get_filtered_value(fir_filter_t *filter);
+    int32_t get_filtered_value(fir_filter_t *filter);
 
 #endif
 /* [] END OF FILE */
